% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gMCPLite-package.R
\docType{package}
\name{AdaptGMCP-package}
\alias{AdaptGMCP}
\alias{AdaptGMCP-package}
\title{AdaptGMCP: Adaptive Weighted Test by using graphs}
\description{
(In Progress)This package provides tools to perform multiple testing based on user inputs.
}
\keyword{internal}
