% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaptGmcpMain.R
\name{adaptGMCP_PC}
\alias{adaptGMCP_PC}
\title{Function to perform Adaptive GMCP by using Combined P-Value(Inverse Normal) by user input P-values}
\usage{
adaptGMCP_PC(
  WI = rep(1/4, 4),
  G = matrix(c(0, 1/3, 1/3, 1/3, 1/3, 0, 1/3, 1/3, 1/3, 1/3, 0, 1/3, 1/3, 1/3, 1/3, 0),
    nrow = 4),
  Threshold = c(0.001, 0.0015, 0.025),
  Correlation = matrix(c(1, 0.5, NA, NA, 0.5, 1, NA, NA, NA, NA, 1, NA, NA, NA, NA, 1),
    nrow = 4),
  W_Norm = sqrt(1:(length(Threshold) - 1)/length(Threshold)),
  Selection = F,
  UpdateStrategy = F
)
}
\arguments{
\item{WI}{Vector of Initial Weights for Global Null(default = \code{rep(1/4,4)})}

\item{G}{Transition Matrix (default = \code{matrix(c(0,1/3,1/3,1/3,  1/3,0,1/3,1/3, 1/3,1/3,0,1/3, 1/3,1/3,1/3,0),nrow = 4)})}

\item{Threshold}{Vector of Critical points to reject H0i in terms of p-values in the interim looks (default = \code{c(0.001,0.0015,0.025)})}

\item{Correlation}{Matrix of correlation between test statistics, NA if the correlation is unknown}

\item{W_Norm}{Weights for inverse normal combining p-value function(default = \code{sqrt(c(1/K, 2/K,..,(K-1)/K)})}
}
\description{
Function to perform Adaptive GMCP by using Combined P-Value(Inverse Normal) by user input P-values
}
